+++ src/models/SnippetBuild.js
@@ -1,0 +1,6 @@
+class SnippetBuild {
+  /** @type {string} **/
+  liquidCode
+}
+
+export default SnippetBuild