+++ src/models/static/Session.js
@@ -14,7 +14,7 @@ class Session {
   static collections
 
   /** @type {string[]} Target Component Name **/
-  static components
+  static componentNames
 
   /** @type {string} Path to external components **/
   static componentsPath